backend/requirements.txt
Django==5.2.2
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework-simplejwt==5.5.0

psycopg[binary]==3.2.9

gunicorn==23.0.0
whitenoise==6.11.0

python-dotenv==1.2.1

openai==2.9.0
google-generativeai==0.8.5
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
proto-plus==1.26.1
protobuf==5.29.5

razorpay==2.0.0

django-dbml==0.10.1
